requirements.txt
setuptools==52.0.0
wheel==0.36.2
gym==0.18.0
matplotlib==3.3.4
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
pyyaml==5.4.1
ray==1.1.0
pickle5==0.0.11
